feature/#21: auth

+++ src/tasks/repository.ts
@@ -1,11 +1,35 @@
 import { pool } from '../database';
 
-const createNewTask = async (title: string, description: string): Promise<iTask | null> => {
+const getAllTasksDB = async (user_id: number): Promise<iTask[] | null> => {
+  try {
+    const sql = 'SELECT * FROM tasks WHERE user_id = $1';
+    const arrOfVal = (await pool.query(sql, [user_id])).rows;
+    if (arrOfVal.length > 0) return arrOfVal;
+    return null;
+  } catch (err) {
+    console.log(`Exception in getTaskById: ${err}`);
+    return null;
+  }
+};
+
+const getTaskById = async (id: number, user_id: number): Promise<iTask | null> => {
+  try {
+    const sql = 'SELECT * FROM tasks WHERE id = $1 AND user_id = $2';
+    const arrOfVal = (await pool.query(sql, [id, user_id])).rows;
+    if (arrOfVal.length > 0) return arrOfVal;
+    return null;
+  } catch (err) {
+    console.log(`Exception in getTaskById: ${err}`);
+    return null;
+  }
+};
+
+const createNewTask = async (title: string, description: string, user_id: number): Promise<iTask | null> => {
   const client = await pool.connect();
   try {
     await client.query('BEGIN');
-    const sql = 'INSERT INTO tasks (title, description) VALUES($1, $2) RETURNING tasks.*';
-    const arrOfVal = (await client.query(sql, [title, description])).rows;
+    const sql = 'INSERT INTO tasks (title, description, user_id) VALUES($1, $2, $3) RETURNING tasks.*';
+    const arrOfVal = (await client.query(sql, [title, description, user_id])).rows;
     await client.query('COMMIT');
     if (arrOfVal.length > 0) return arrOfVal;
     return null;
@@ -18,35 +42,12 @@ const createNewTask = async (title: string, description: string): Promise<iTask 
   }
 };
 
-const getAllTasksDB = async (): Promise<iTask[] | null> => {
-  try {
-    const arrOfVal = (await pool.query(`SELECT * FROM tasks`)).rows;
-    if (arrOfVal.length > 0) return arrOfVal;
-    return null;
-  } catch (err) {
-    console.log(`Exception in getTaskById: ${err}`);
-    return null;
-  }
-};
-
-const getTaskById = async (id: number): Promise<iTask | null> => {
-  try {
-    const sql = 'SELECT * FROM tasks WHERE id = $1';
-    const arrOfVal = (await pool.query(sql, [id])).rows;
-    if (arrOfVal.length > 0) return arrOfVal;
-    return null;
-  } catch (err) {
-    console.log(`Exception in getTaskById: ${err}`);
-    return null;
-  }
-};
-
-const updateTaskById = async (id: number, title: string, description: string): Promise<iTask | null> => {
+const updateTaskById = async (id: number, title: string, description: string, user_id: number): Promise<iTask | null> => {
   const client = await pool.connect();
   try {
     await client.query('BEGIN');
-    const sql = 'UPDATE tasks SET title = $1, description = $2 WHERE id = $3 RETURNING tasks.*';
-    const arrOfVal = (await client.query(sql, [title, description, id])).rows;
+    const sql = 'UPDATE tasks SET title = $1, description = $2 WHERE id = $3 AND user_id = $4 RETURNING tasks.*';
+    const arrOfVal = (await client.query(sql, [title, description, id, user_id])).rows;
     await client.query('COMMIT');
     if (arrOfVal.length > 0) return arrOfVal;
     return null;
@@ -59,12 +60,12 @@ const updateTaskById = async (id: number, title: string, description: string): P
   }
 };
 
-const deleteTaskById = async (id: number): Promise<iTask | null> => {
+const deleteTaskById = async (id: number, user_id: number): Promise<iTask | null> => {
   const client = await pool.connect();
   try {
     await client.query('BEGIN');
-    const sql = 'DELETE FROM tasks WHERE id = $1 RETURNING tasks.*';
-    const arrOfVal = (await client.query(sql, [id])).rows;
+    const sql = 'DELETE FROM tasks WHERE id = $1 AND user_id = $2 RETURNING tasks.*';
+    const arrOfVal = (await client.query(sql, [id, user_id])).rows;
     await client.query('COMMIT');
     if (arrOfVal.length > 0) return arrOfVal;
     return null;
